+++ libEM/sparx/util_sparx.h
@@ -640,7 +640,7 @@ class FakeKaiserBessel : public KaiserBessel {
 
 	static vector<float> FCross_multiref(EMData* frobj, EMData* frings, int psi_start, int psi_step);
 
-	static void multiply_rows( EMData* rings, const vector<int>& bckgnoise, int nb );
+	static void multiply_rows( EMData* rings, const vector<float>& bckgnoise, int nb );
 
 	/** Single Precision Fourier Transform for a set of rings */
 	static void  Frngs(EMData* circ, vector<int> numr);
